# docs/rest/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentation for the Project Odeon backend",
    "version" : "v2",
    "title" : "Swagger documentation for the Project Odeon backend"
  },
  "host" : "jscpeterson.com",
  "basePath" : "/rest/",
  "schemes" : [ "https" ],
  "paths" : {
    "/film_locations" : {
      "get" : {
        "summary" : "Gets a list of film locations",
        "description" : "Gets a list of film locations.",
        "operationId" : "list",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FilmLocation"
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Posts a film location.",
        "description" : "",
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/FilmLocation"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FilmLocation"
            }
          }
        }
      },
      "patch" : {
        "summary" : "Patches a single film location",
        "description" : "Patches a film location. All data will be overwritten - current fields to stay the same must be included as well.",
        "operationId" : "patch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/FilmLocation"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/film_locations/{filmLocationId}" : {
      "get" : {
        "summary" : "Gets a single film location",
        "description" : "Gets a single film location.",
        "operationId" : "get",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FilmLocation"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes a film location",
        "description" : "Deletes a film location",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    },
    "/film_locations/{filmLocationId}/images" : {
      "get" : {
        "summary" : "Gets all images on a FilmLocation.",
        "description" : "",
        "operationId" : "getImages",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Image"
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Posts an image to a film location.",
        "description" : "",
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Image"
          }
        }, {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Image"
            }
          }
        }
      }
    },
    "/film_locations/{filmLocationId}/images/" : {
      "patch" : {
        "summary" : "Patches an image. All data will be overwritten.",
        "description" : "",
        "operationId" : "patch",
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Image"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/film_locations/{filmLocationId}/images/{imageId}" : {
      "get" : {
        "summary" : "Gets all images on a FilmLocation.",
        "description" : "",
        "operationId" : "getImage",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "imageId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Image"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes an image on a film location.",
        "description" : "",
        "operationId" : "deleteImage",
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "imageId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    },
    "/film_locations/{filmLocationId}/user_comments" : {
      "get" : {
        "summary" : "Gets all user comments on a location.",
        "description" : "",
        "operationId" : "getUserComments",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserComment"
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Posts a user comment to a location.",
        "description" : "",
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserComment"
          }
        }, {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserComment"
            }
          }
        }
      },
      "patch" : {
        "summary" : "Patches a UserComment on a location. All data to stay the same must be included.",
        "description" : "",
        "operationId" : "patch",
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserComment"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/film_locations/{filmLocationId}/user_comments/{userCommentId}" : {
      "get" : {
        "summary" : "Gets a single UserComment from a location.",
        "description" : "",
        "operationId" : "getUserComment",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "userCommentId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserComment"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Deletes a UserComment from a location.",
        "description" : "",
        "operationId" : "deleteUserComment",
        "parameters" : [ {
          "name" : "filmLocationId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "userCommentId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          }
        }
      }
    },
...
  },
  "definitions" : {
    "FilmLocation" : {
      "type" : "object",
      "properties" : {
        "userId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "user" : {
          "$ref" : "#/definitions/GoogleUser"
        },
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "longCoordinate" : {
          "type" : "number",
          "format" : "double"
        },
        "latCoordinate" : {
          "type" : "number",
          "format" : "double"
        },
        "siteName" : {
          "type" : "string"
        },
        "imdbId" : {
          "type" : "string"
        },
        "productionId" : {
          "type" : "string"
        },
        "production" : {
          "$ref" : "#/definitions/Production"
        },
        "address" : {
          "type" : "string"
        },
        "shootDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "originalDetails" : {
          "type" : "string"
        },
        "approved" : {
          "type" : "boolean"
        },
        "googleId" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "GoogleUser" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "googleId" : {
          "type" : "string"
        },
        "googleName" : {
          "type" : "string"
        },
        "gmailAddress" : {
          "type" : "string"
        },
        "banned" : {
          "type" : "boolean"
        },
        "banReason" : {
          "type" : "string"
        },
        "userRole" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "Image" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "filmLocation" : {
          "$ref" : "#/definitions/FilmLocation"
        },
        "description" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "userId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "user" : {
          "$ref" : "#/definitions/GoogleUser"
        },
        "googleId" : {
          "type" : "string"
        },
        "approved" : {
          "type" : "boolean"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
    "Production" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "imdbId" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "releaseYear" : {
          "type" : "string"
        },
        "plot" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    },
...
    "UserComment" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "filmLocation" : {
          "$ref" : "#/definitions/FilmLocation"
        },
        "text" : {
          "type" : "string"
        },
        "userId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "user" : {
          "$ref" : "#/definitions/GoogleUser"
        },
        "googleId" : {
          "type" : "string"
        },
        "approved" : {
          "type" : "boolean"
        },
        "href" : {
          "type" : "string",
          "format" : "uri"
        }
      }
    }
  }
}